third_party/blink/web_tests/platform/generic/html/document_metadata/style-disabled-expected.txt

Test the disabled property on a style element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS styleElement.disabled is false
PASS window.getComputedStyle(testElement).color is "rgb(0, 128, 0)"
PASS styleElement.disabled is true
PASS window.getComputedStyle(testElement).color is "rgb(255, 0, 0)"
PASS styleElement.sheet.disabled is true
PASS styleElement.sheet.disabled is false
PASS styleElement.disabled is false
PASS window.getComputedStyle(testElement).color is "rgb(0, 128, 0)"
PASS newStyleElement.disabled is false
PASS newStyleElement.disabled is false
PASS otherStyle.disabled is false
PASS otherStyle.disabled is false
PASS successfullyParsed is true

TEST COMPLETE
